Make cleaner separation between parser and main.
Currently the main interface creates CLArgs to parser arguments, which (depending on input) get passed down into the vroom machinery.
For the most part this is done cleanly in such a way that you don't need CLArgs to use any machinery, with input parser being the only exception.

I propose we make the interface cleaner by keeping CLArgs contain to the main file and just passing it's attribute along explicitly.